Flushing of serial buffer before sending
S: Added error handling (exceptions) for the detection of the devices and checking if they are plugged in.

S: Added a buffer clearing in the "sending functions" before sending data to the boards.

R: CRC function call was not properly integrated. Coud not have caused errors, but did not double check before sending data. Multiple "sending" tries are already implemented, so this would just make a timing difference.